+++ source/Examples/Resources/SqlAG/1-CreateAvailabilityGroup.ps1
@@ -29,7 +29,7 @@ Configuration Example
         }
 
         # Add the required permissions to the cluster service login
-        SqlServerPermission 'AddNTServiceClusSvcPermissions'
+        SqlPermission 'AddNTServiceClusSvcPermissions'
         {
             DependsOn            = '[SqlLogin]AddNTServiceClusSvc'
             Ensure               = 'Present'
@@ -72,7 +72,7 @@ Configuration Example
             InstanceName         = 'MSSQLSERVER'
             ServerName           = $Node.NodeName
 
-            DependsOn            = '[SqlAlwaysOnService]EnableHADR', '[SqlEndpoint]HADREndpoint', '[SqlServerPermission]AddNTServiceClusSvcPermissions'
+            DependsOn            = '[SqlAlwaysOnService]EnableHADR', '[SqlEndpoint]HADREndpoint', '[SqlPermission]AddNTServiceClusSvcPermissions'
 
             PsDscRunAsCredential = $SqlAdministratorCredential
         }